Move the Backend driver's Synchronizers class to States.h
